Add a GtkRadioGroup::active-value property
This is only supported for radio groups which contain radio actions,
and in that case is just the ::string-value property of the active item.
The property is readable and writable. Setting it has the effect of
changing the active item, if any of the items has a matching string-value.

This property is intended for easy binding to settings.