Clarify documentation header about GTK+ 3 vs 2
I think it's confusing for a lot the developers out there who
may not even be aware of GTK+ 3 coming, if suddenly GTK+ 3 becomes
the "stable" version of "gtk" on library.gnome.org.  It may
not even be feasible for them to port to GTK+3 if it's not
shipped in the operating systems they're targeting (for example,
RHEL 6).

Since practically speaking, we expect people to consume GTK+ 2 for
several years at least, redirect these people to the right pages.

(I didn't attempt to explain the differences between the libraries
 here, but hopefully the major version difference is enough of a hint)

As a side effect, this makes the generated HTML look better; previously
it looked rather crappy, since the "for GTK &version;" was totally
offset and in a different group from the documentation title.