Bug 643270: VS 2010 Project Files
These are the VS2010 project files to compile GDK and GTK+
with the Win32 backend, with a README.txt file to briefly
explain the steps to build this.

The main GDK and GTK+ project/filter files are templates
that are expanded during "make dist" so that the source
file listings in them will be up-to-date as far as
possible, simplifying maintenance.

Any comments and suggestions are welcome.